# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Petstore - OpenAPI 3.0",
    "description": "",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get all users",
        "description": "Get full list of all users",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation exception"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "write:users",
              "read:users"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add a new user",
        "description": "Add a new user",
        "operationId": "addUser",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "user": {
                  "$ref": "#/components/examples/User"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "write:users",
              "read:users"
            ]
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get user by id",
        "description": "Get full data of a user",
        "operationId": "getUserById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the user to get"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation exception"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "write:users",
              "read:users"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user by id",
        "description": "Update the datas of a user",
        "operationId": "UpdateUserById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the user to update"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "car": {
                  "$ref": "#/components/examples/User"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Car not found"
          },
          "422": {
            "description": "Validation exception"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "write:users",
              "read:users"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete user by id",
        "description": "Delete the data of a user",
        "operationId": "DeleteUserById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the user to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation exception"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "write:users",
              "read:users"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "required": [
          "name",
          "email",
          "password",
          "address",
          "phoneNumber",
          "role"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@domain.com"
          },
          "password": {
            "type": "string",
            "example": "YourPa$$w0rd"
          },
          "address": {
            "type": "string",
            "example": "My City Name"
          },
          "phoneNumber": {
            "type": "string",
            "example": "087654321234"
          },
          "role": {
            "type": "string",
            "example": "User"
          },
          "avatar": {
            "type": "string",
            "example": "url://something"
          },
          "driverLicense": {
            "type": "string",
            "example": "06721174761"
          },
          "birthdate": {
            "type": "string",
            "example": "2018-03-01T00:00:00.000Z"
          }
        }
      },
...
    },
...
    "examples": {
...
      "User": {
        "value": {
          "name": "John Doe",
          "email": "john.doe@domain.com",
          "password": "YourPa$$w0rd",
          "address": "My City Name",
          "phoneNumber": "087654321234",
          "role": "User",
          "avatar": "url://something",
          "driverLicense": "06721174761",
          "birthdate": "2018-02-29"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}